Hide balloon notification after 6 seconds.
